First Working Prototype!
- it works mostly
- unit tests coming later
