--- src/main/java/SYSC6/Project/employee/Employee.java
@@ -1,17 +1,0 @@
-package SYSC6.Project.employee;
-
-import SYSC6.Project.RoleType;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class Employee {
-
-    private Integer id;
-    private String name;
-    private String role;
-
-}
