+++ src/main/java/SYSC6/Project/user/UserUtil.java
@@ -1,0 +1,9 @@
+package SYSC6.Project.user;
+
+public class UserUtil {
+
+    public static User setRoleToUpgradeRole(User user){
+        user.setRole(user.getRole().getUpgradeRole().get());
+        return user;
+    }
+}
