+++ src/main/java/SYSC6/Project/user/userManagement/UserPOJO.java
@@ -1,14 +1,18 @@
-package SYSC6.Project.employee;
+package SYSC6.Project.user.userManagement;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+/**
+ * Lightweight user object for use in front end elements
+ */
 @Getter
 @Setter
 @NoArgsConstructor
 public class UserPOJO {
 
+    private Long id;
     private String username;
     private String role;
