+++ src/main/resources/templates/admin_portal.html
@@ -15,7 +15,7 @@
 <form action="#" th:action="@{/LogOut}" method="post">
     <button type="submit" class="float">Log out</button>
 </form>
-<form action="#" th:action="@{/view_users}" method="post">
+<form action="#" th:action="@{/view_users}" method="get">
     <button type="submit" class="float">View list of Users</button>
 </form>
