+++ src/main/resources/templates/view_users.html
@@ -13,7 +13,7 @@
 <!-- Navigation -->
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark static-top">
     <div class="container">
-        <a class="navbar-brand" href="/">Thymeleaf - Bootstrap DataTable</a>
+        <a class="navbar-brand" href="/">Admin Portal - User Management</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive"
                 aria-controls="navbarResponsive"
                 aria-expanded="false" aria-label="Toggle navigation">
@@ -22,18 +22,15 @@
         <div class="collapse navbar-collapse" id="navbarResponsive">
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item active">
-                    <a class="nav-link" href="#">Home
+                    <a class="nav-link" href="#" >User Management
                         <span class="sr-only">(current)</span>
                     </a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="#">About</a>
+                    <a class="nav-link" href="http://localhost:8080/" th:action="@{/LogOut}">Admin-Home</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="#">Services</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="#">Contact</a>
+                    <a class="nav-link" href="/" th:action="@{/LogOut}">Logout</a>
                 </li>
             </ul>
         </div>
@@ -44,11 +41,11 @@
     <div class="row">
         <div class="col-lg-10 mt-5 mb-5">
             <h3>Datatable using list of Objects in data parameter</h3>
-            <table id="example" class="table table-bordered table-responsive" style="width: 100%">
+            <table id="users" class="table table-bordered table-responsive" style="width: 100%">
                 <thead>
                 <tr>
                     <th style="width:50%">Username</th>
-                    <th style="width:100%">Role</th>
+                    <th style="width:70%">Role</th>
                     <th style="width:100%">Interactions</th>
                 </tr>
                 </thead>
@@ -62,32 +59,6 @@
             </table>
         </div>
     </div>
-
-    <!--
-    <div class="row">
-        <div class="col-lg-10 mt-5 mb-5">
-            <h3>Datatable using list of Arrays in data parameter</h3>
-            <table id="exampleArray" class="table table-bordered table-responsive" style="width: 100%">
-                <thead>
-                <tr>
-                    <th style="width:20%">Id</th>
-                    <th style="width:30%">Name</th>
-                    <th style="width:30%">Role</th>
-                    <th style="width:20%">Interactions</th>
-                </tr>
-                </thead>
-                <tfoot>
-                <tr>
-                    <th>Id</th>
-                    <th>Name</th>
-                    <th>Role</th>
-                    <th>Interactions</th>
-                </tr>
-                </tfoot>
-            </table>
-        </div>
-    </div>
-    -->
 </div>
 
 
@@ -97,11 +68,12 @@
 <script th:src="@{datatable/datatables.js}"></script>
 
 <script>
-    var table = $('#example').DataTable({
-        "processing": true,
+
+    var table = $('#users').DataTable({
+        "processing": false,
         "serverSide": true,
         "ajax": {
-            "url": "/employees/users",
+            "url": "/rest/api/usertable",
             "type": "POST",
             "dataType": "json",
             "contentType": "application/json",
@@ -112,31 +84,32 @@
         "columns": [
             {"data": "username"},
             {"data": "role"},
-            {"data": null, "defaultContent": "<button>Toggle Role</button>"}
+            {"data": null, "defaultContent": "<button>Upgrade Role</button>"}
         ],
     });
 
-    $('#example tbody').on( 'click', 'button', function () {
+    $('#users tbody').on( 'click', 'button', function () {
         let data = table.row( $(this).parents('tr') ).data();
-        console.log( data ); //TODO DEBUG
-        alert( data["username"] +"'s role is: "+ data["role"] );
+        toggleRoleFunction(data["id"]);
+        table.ajax.reload();
     } );
 
-    /*
-    $('#exampleArray').DataTable({
-        "processing": true,
-        "serverSide": true,
-        "ajax": {
-            "url": "/employees/array",
-            "type": "POST",
-            "dataType": "json",
-            "contentType": "application/json",
-            "data": function (d) {
-                return JSON.stringify(d);
+    setInterval(function() {
+        table.ajax.reload();
+    }, 1000);
+
+    function toggleRoleFunction(id){
+        $.ajax({
+            type: "PUT",
+            url: "http://localhost:8080/rest/api/user/upgrade/"+ id,
+            contentType : "application/json",
+            success: function (d) {
+            },
+            error: function (d) {
+                alert("Error Upgrading User");
             }
-        }
-    });
-    */
+        });
+    }
 </script>
 
 </body>
